feat: implement Clerk js for auth; signin, signup, logout, sessions,
